Add Toasts for adding and refreshing Toons
Figuring out where to place relevant Toasts was challenging since the
API calls are done synchronously in the background. ToonDb.addToon(Toon)
was modified to return a boolean signifying whether the Toon was added
or updated, which was used in the new TLF.addProcessedToon(Toon) method
to determine which String to display. The Adapter and Holder swapped
places in TLF just to improve flow readability.